AsyncAbstractResponse: Fix packet buffer lifetime
Packet buffers passed to lwip are required to remain allocated until
acked, in case they need to be retransmitted.  Replace the stack-scoped
local packet buffer in AsyncAbstractResponse::_ack with an object-
scoped one.  We apply an allocator trick to the vector to speed up its
usage with POD types where we don't care if they're zero-initialized.